Beta version of MVTX service barrel
This version contains three concentric cylinders of copper, water and finally polyethylene. The layers begin at r = 8.5cm and z = -25cm (the z-gap will likely be filled by the FPC extensions). Each layer is a solid ring, accounting for the total material contribution from 48 staves.

The number of service layers can be easily configured by setting the number of layers, layer start position and number of staves supplied by that layer. The calculations and placements are then done by the rest of the macro

Additionally, this macro contains code with parameters for all the individual cables and tubing but is currently commented out. For example, this means there is code for the number of ground wires with the inner and outer radii of the wire that each stave requires. A function can be written at a later date to correctly produce each service bundle when this is defined, if required.