KFParticle macro added
This PR adds the KFParticle macro to perform Upsilon, D0 and Lc reco

The three reconstructions were tested and work without crashing although the Lc generator is still a work in progress. This will appear in a later PR to core software.

There are some base cuts which are realistic but might be tight for code testing so they can be loosened (or set to effectively no cut). The nTuple output is disabled by default while the DST output is enabled. Truth matching is also disabled by default, only enable it if you know there are truth variables. KFParticle will exit your run rather than skip the event if this is the case